Revert implementation of some edit dialogs
Paths begin with `/src/`.

In commit be72c59, the "Edit Dos and Don'ts" and "Edit Requirements"
dialogs were implemented.
The pull request was not merged, and another implementation was pushed
to the repo instead.
This commit reverts the implementation of said dialogs.

Combined a function in `controller/campaignController.ts` with one in
`service/campaignServices.ts`.